Rename ErrorMassage component directory
